Agent2_Free/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
ddgs>=0.1.0
python-dotenv>=1.0.0
huggingface_hub>=0.23.0
